Remove thumbprint and add static parser
Removes the Thumbprint property. Instead, the Subject property will be used to define the subject, distinguished name or thumbprint of a store certificate.

Also adds a property that specifies if the private key is required when loading a store certificate, and adds a static CreateFromPath method that can parse a file system or certificate store path.